fix(procurement): critical bugs found via real local DB testing
Two bugs discovered by running the tests against an isolated PostgreSQL 16
instance (localhost:55432, fresh cluster in scratchpad):

1. pg_advisory_xact_lock(bigint, bigint) does NOT exist in PostgreSQL.
   The two-arg form requires (integer, integer). Changed all 11 call sites
   across migrations 2, 3, 4, 5 to use the correct signature:
   pg_advisory_xact_lock(hashtext('proc_requisitions'), p_req_id::INT)

2. Guard trigger functions were declared SECURITY DEFINER, which caused
   current_user inside them to be the definer (postgres) instead of the
   caller — completely defeating the current_user check. Added new
   Migration 6 that redefines the three guard functions
   (proc_req_update_guard, proc_req_items_write_guard,
   proc_po_creation_guard) as SECURITY INVOKER so current_user reflects
   the actual DML caller.

Also updated proc-approval-tests.sql:
- Tests 9/10 now use PERFORM _test_as(99003) + SET LOCAL ROLE authenticated
  to simulate an authenticated client's direct UPDATE (which is what the
  guards are designed to catch).
- Added TEST Vendor seed (id=99001) so tests 11/12/25 don't SKIP.

Local run results (real PostgreSQL 16, isolated cluster):
- 40 PASS (all sub-tests), 1 SKIPPED (concurrent — separate script)
- Includes all G2 notification tests (30-34) and G6 audit tests (27-29)
- Snapshot immutability + AMBIGUOUS_APPROVAL_RULES + RULE_IN_USE verified

+++ proc-approval-2-hardening.sql
@@ -243,10 +243,8 @@ BEGIN
   END IF;
 
   -- (1) قفل الطلب على مستوى المعاملة (Advisory Lock)
-  PERFORM pg_advisory_xact_lock(
-    hashtext('proc_requisitions')::BIGINT,
-    p_req_id
-  );
+  --     نستخدم توقيع (integer, integer) لأن (bigint, bigint) غير موجود.
+  PERFORM pg_advisory_xact_lock(hashtext('proc_requisitions'), p_req_id::INT);
 
   SELECT * INTO v_req FROM proc_requisitions WHERE id = p_req_id FOR UPDATE;
   IF NOT FOUND THEN RAISE EXCEPTION 'REQ_NOT_FOUND: الطلب غير موجود'; END IF;
@@ -366,7 +364,7 @@ BEGIN
   END IF;
 
   -- Advisory Lock على الطلب — أي عملية اعتماد/رفض على نفس الطلب تنتظر
-  PERFORM pg_advisory_xact_lock(hashtext('proc_requisitions')::BIGINT, v_req_id);
+  PERFORM pg_advisory_xact_lock(hashtext('proc_requisitions'), v_req_id::INT);
 
   -- الآن اقفل الخطوة والطلب معًا وأعد قراءتهما
   SELECT * INTO v_step FROM proc_requisition_approvals WHERE id = p_step_id FOR UPDATE;
@@ -458,7 +456,7 @@ BEGIN
   IF v_req_id IS NULL THEN
     RAISE EXCEPTION 'STEP_NOT_FOUND: خطوة الاعتماد غير موجودة';
   END IF;
-  PERFORM pg_advisory_xact_lock(hashtext('proc_requisitions')::BIGINT, v_req_id);
+  PERFORM pg_advisory_xact_lock(hashtext('proc_requisitions'), v_req_id::INT);
 
   SELECT * INTO v_step FROM proc_requisition_approvals WHERE id = p_step_id FOR UPDATE;
   IF v_step.status <> 'pending' THEN
@@ -518,7 +516,7 @@ BEGIN
     RAISE EXCEPTION 'USER_INACTIVE: المستخدم غير نشط';
   END IF;
 
-  PERFORM pg_advisory_xact_lock(hashtext('proc_requisitions')::BIGINT, p_req_id);
+  PERFORM pg_advisory_xact_lock(hashtext('proc_requisitions'), p_req_id::INT);
 
   SELECT * INTO v_req FROM proc_requisitions WHERE id = p_req_id FOR UPDATE;
   IF NOT FOUND THEN RAISE EXCEPTION 'REQ_NOT_FOUND: الطلب غير موجود'; END IF;
@@ -594,7 +592,7 @@ BEGIN
     RAISE EXCEPTION 'PERMISSION_DENIED: يشترط دور مدير مشتريات/مالية';
   END IF;
 
-  PERFORM pg_advisory_xact_lock(hashtext('proc_requisitions')::BIGINT, p_req_id);
+  PERFORM pg_advisory_xact_lock(hashtext('proc_requisitions'), p_req_id::INT);
 
   SELECT * INTO v_req FROM proc_requisitions WHERE id = p_req_id FOR UPDATE;
   IF NOT FOUND THEN RAISE EXCEPTION 'REQ_NOT_FOUND: الطلب غير موجود'; END IF;
