+++ proc-approval-5-notifications-audit.sql
@@ -286,7 +286,7 @@ BEGIN
   SELECT COALESCE(is_active, TRUE) INTO v_caller_ok FROM users WHERE id = v_caller;
   IF NOT v_caller_ok THEN RAISE EXCEPTION 'USER_INACTIVE: المستخدم غير نشط'; END IF;
 
-  PERFORM pg_advisory_xact_lock(hashtext('proc_requisitions')::BIGINT, p_req_id);
+  PERFORM pg_advisory_xact_lock(hashtext('proc_requisitions'), p_req_id::INT);
 
   SELECT * INTO v_req FROM proc_requisitions WHERE id = p_req_id FOR UPDATE;
   IF NOT FOUND THEN RAISE EXCEPTION 'REQ_NOT_FOUND: الطلب غير موجود'; END IF;
@@ -444,7 +444,7 @@ BEGIN
   SELECT requisition_id INTO v_req_id FROM proc_requisition_approvals WHERE id = p_step_id;
   IF v_req_id IS NULL THEN RAISE EXCEPTION 'STEP_NOT_FOUND: خطوة الاعتماد غير موجودة'; END IF;
 
-  PERFORM pg_advisory_xact_lock(hashtext('proc_requisitions')::BIGINT, v_req_id);
+  PERFORM pg_advisory_xact_lock(hashtext('proc_requisitions'), v_req_id::INT);
 
   SELECT * INTO v_step FROM proc_requisition_approvals WHERE id = p_step_id FOR UPDATE;
   IF v_step.status <> 'pending' THEN
@@ -545,7 +545,7 @@ BEGIN
 
   SELECT requisition_id INTO v_req_id FROM proc_requisition_approvals WHERE id = p_step_id;
   IF v_req_id IS NULL THEN RAISE EXCEPTION 'STEP_NOT_FOUND: خطوة الاعتماد غير موجودة'; END IF;
-  PERFORM pg_advisory_xact_lock(hashtext('proc_requisitions')::BIGINT, v_req_id);
+  PERFORM pg_advisory_xact_lock(hashtext('proc_requisitions'), v_req_id::INT);
 
   SELECT * INTO v_step FROM proc_requisition_approvals WHERE id = p_step_id FOR UPDATE;
   IF v_step.status <> 'pending' THEN
